# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MirrorJobInput {
  /** HTTPS or HTTP URL for a site the caller owns or is authorized to archive */
  url: string;
  /**
     * @minimum 1
     * @maximum 1000
     */
  maxPages?: number;
  /**
     * @minimum 0
     * @maximum 5000
     */
  requestDelayMs?: number;
  respectRobotsTxt?: boolean;
  /**
     * @minimum 0
     * @maximum 10
     */
  maxDepth?: number;
  includeAssets?: boolean;
  /**
     * @minLength 1
     * @maxLength 200
     */
  pathPrefix?: string;
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 200
     */
  excludePaths?: string[];
  /**
     * Wall-clock limit for the whole crawl, in milliseconds.
     * @minimum 60000
     * @maximum 3600000
     */
  timeoutMs?: number;
  /**
     * Overall archive size cap, in bytes.
     * @minimum 1048576
     * @maximum 2147483648
     */
  maxTotalBytes?: number;
}

export type MirrorJobStatus = typeof MirrorJobStatus[keyof typeof MirrorJobStatus];


export const MirrorJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  completed_with_warnings: 'completed_with_warnings',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type MirrorJobProgressPhase = typeof MirrorJobProgressPhase[keyof typeof MirrorJobProgressPhase];


export const MirrorJobProgressPhase = {
  queued: 'queued',
  discovering: 'discovering',
  saving: 'saving',
  downloading_assets: 'downloading_assets',
  rewriting: 'rewriting',
  packaging: 'packaging',
} as const;

export interface MirrorJob {
  id: string;
  url: string;
  status: MirrorJobStatus;
  pagesFound: number;
  pagesDownloaded: number;
  pagesSkipped: number;
  pagesFailed: number;
  assetsDownloaded: number;
  assetsSkipped: number;
  assetsFailed: number;
  bytesDownloaded: number;
  maxPages: number;
  requestDelayMs: number;
  respectRobotsTxt: boolean;
  maxDepth?: number;
  includeAssets?: boolean;
  pathPrefix?: string;
  excludePaths?: string[];
  timeoutMs?: number;
  maxTotalBytes?: number;
  /** @nullable */
  currentUrl: string | null;
  progressPhase: MirrorJobProgressPhase;
  /** @nullable */
  message: string | null;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
}

export interface MirrorJobList {
  jobs: MirrorJob[];
}

export interface ErrorResponse {
  error: string;
}

export type ListMirrorJobsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};