# lib/api-zod/src/generated/types/mirrorJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MirrorJobProgressPhase } from './mirrorJobProgressPhase';
import type { MirrorJobStatus } from './mirrorJobStatus';

export interface MirrorJob {
  id: string;
  url: string;
  status: MirrorJobStatus;
  pagesFound: number;
  pagesDownloaded: number;
  pagesSkipped: number;
  pagesFailed: number;
  assetsDownloaded: number;
  assetsSkipped: number;
  assetsFailed: number;
  bytesDownloaded: number;
  maxPages: number;
  requestDelayMs: number;
  respectRobotsTxt: boolean;
  maxDepth?: number;
  includeAssets?: boolean;
  pathPrefix?: string;
  excludePaths?: string[];
  timeoutMs?: number;
  maxTotalBytes?: number;
  /** @nullable */
  currentUrl: string | null;
  progressPhase: MirrorJobProgressPhase;
  /** @nullable */
  message: string | null;
  createdAt: Date;
  /** @nullable */
  completedAt: string | null;
}