# lib/api-zod/src/generated/types/mirrorJobInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MirrorJobInput {
  /** HTTPS or HTTP URL for a site the caller owns or is authorized to archive */
  url: string;
  /**
     * @minimum 1
     * @maximum 1000
     */
  maxPages?: number;
  /**
     * @minimum 0
     * @maximum 5000
     */
  requestDelayMs?: number;
  respectRobotsTxt?: boolean;
  /**
     * @minimum 0
     * @maximum 10
     */
  maxDepth?: number;
  includeAssets?: boolean;
  /**
     * @minLength 1
     * @maxLength 200
     */
  pathPrefix?: string;
  /**
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 200
     */
  excludePaths?: string[];
  /**
     * Wall-clock limit for the whole crawl, in milliseconds.
     * @minimum 60000
     * @maximum 3600000
     */
  timeoutMs?: number;
  /**
     * Overall archive size cap, in bytes.
     * @minimum 1048576
     * @maximum 2147483648
     */
  maxTotalBytes?: number;
}